+++ CHANGELOG.rst
@@ -1,10 +1,11 @@
 Change Log
 =============
 
-[4.12.0] - 2023-08-06
+[4.12.0] - 2023-08-08
 ----------------------
-- [CHANGED] added IndBinRsrInvest option to relax reservoir investment decisions. This is needed to keep compatibility with previous cases
-- [CHANGED] added production function to hydropower plants to be modeled in water units instead of energy units. This is needed to keep compatibility with previous cases
+- [CHANGED] added eTrbReserveUpIfEnergy, eTrbReserveDwIfEnergy, ePmpReserveUpIfEnergy, and ePmpReserveDwIfEnergy constraints
+- [CHANGED] added IndBinRsrInvest in Option file to relax reservoir investment decisions. This is needed to keep compatibility with previous cases
+- [CHANGED] added production function to hydropower plants in Generation file to be modeled in water units instead of energy units. This is needed to keep compatibility with previous cases
 - [CHANGED] added dictionaries of hydro basin topology in water units (Dict_Reservoir, Dict_ReservoirToHydro, Dict_HydroToReservoir, Dict_ReservoirToPumpedHydro, Dict_PumpedHydroToReservoir, Dict_ReservoirToReservoir)
 - [CHANGED] added data for water hydro inflows and outflows (Data_HydroInflows, Data_HydroOutflows)
 - [CHANGED] added data for reservoirs (Data_Reservoir, Data_VariableMaxVolume, oT_Data_VariableMinVolume)